requirements.txt
torch==1.31.1
torchvision == 0.14.1
pillow
matplotlib